mdss: fb: fix ion fbmem release sequence
In recovery mode and charger only mode, a race
condition is sometimes seen in the shutdown sequence.
This happens in the mdss_release where the ion buffers
used for FB are freed before doing the BLANK which
results in MDP accessing the buffers even after it
is freed leading to iommu pagefaults.

To resolve this, free the buffers only after the BLANK
operation is completed.

CRs-Fixed: 750233

Change-Id: I347e98f933bfd9372ce45b007b8f737b31e8479d
Signed-off-by: Ramakant Singh <ramaka@codeaurora.org>
Signed-off-by: Nirmal Abraham <nabrah@codeaurora.org>